Added memory_cache_shared to squid.conf. Report whether mem_cache is shared.
Allow the user to explicitly disable shared memory caching in SMP mode.

Added YesNoNone class to allow Squid to compute the default value of a boolean
option (based on other options) instead of using a hard-coded default. The
class is used to automatically enable shared memory caching in SMP
environments (and disable it in non-SMP environments) unless the user says
otherwise.

This needs more work to reduce the number of shared memory pages if shared
memory caching is disabled (the pages may still be needed for SMP I/O).