SourceLayout: separate HttpParser from HttpMsg and HttpRequest files
Paving the way for the parser to be used on replies or outside of HttpRequest.

Also: shuffle request-line offsets into a sub-struct.
Preparation for parsing other segments such as mime headers and reply
status-line we need the segments data offsets to be kept separate.

There are no logic changes in this.

Also adds a stub file for event.cc