Default to vhost for accelerator mode (reverse proxy)
Defaults to match HTTP requirements, and Host awareness is a
rather strong HTTP/1.1 requirement. The default in HTTP/1.1 is to read
the Host header (unless URL is absolute) but a server MAY optionally
ignore the Host header if desired.

The option no-vhost is provided to disable this behaviour if necessary.