Bug 3377: assertion failed: store.cc:885: "store_status == STORE_PENDING"
The StoreEntry::write in the case of an empty write, calls the StoreEntry 
handlers. It is possible one of these handlers will change the state of the
store entry or abort it. The next call of the StoreEntry::write will cause
this assertion.

The block of code which calls the StoreSntry handlers in the case of an empty
write, added to allow forward http headers to the client even if no body data
arrived yet (bug 1750). There is not need for this part of code in the latest
squid releases, so it is safe to be removed.