Check SSL server certificate on reconnect
In the case where the squid-to-server connection closed and we re-connect to 
server we need to check if the server certificate changed.

Moreover we need to check stored certificates with the web server certificate
to handle the case the web server certificate is updated. In this case a new
certificate must generated to mimic the updated server certificate fields