Retry requests that failed due to a persistent connection race instead of replying with ERR_ZERO_SIZE_OBJECT "Bad Gateway".
Trunk r12050 code contains a portion of these changes. The rest is code
to re-pin a bump-server-first connection if a bumped pinned pconn fails
due to a race. We use a new canRePin flag to mark connections that can
be repinned and assume that pinned connections unrelated to SslBump
cannot be repinned.