Certificate for error messages
If a certificate verification error is honored by Squid (i.e., a Squid error 
page is returned to the user), the user gets a browser warning and then, after
ignoring the warning, the squid error message (about the same kind of problem). This happens because the fake certificate mimics details of the true server  
certificate, which is "broken".

When Squid serves its own error page it is no longer trying to hide its presencefrom the user. Thus, it is probably better to serve all Squid-generated SSL 
handshake and validation error responses with a trusted certificate containing 
minimal details, like we already do for the DNS_FAIL errors.

This patch fixing the above problem.

NOTE: The fix will not work for domain mismatch errors. We are checking
for domain mismatch after we have establish the SSL connection with the web 
client , and after we got the http request, because we need the HTTP request
Host header to check for SSL domain mismatch errors.