Bug fix: ssl_crtd crashes when accessing HTTPS sites with a domain name exceeding 64 characters
Squid tries to generate a certificate for long domain names, which is not
possible.                                                                   
According to RFC 5280 (Section A.1), the common name length in a certificate 
can be at most 64 characters. Therefore it is not possible to generate a valid 
certificate with the above domain name as common name.
This patch does not allow use of common names longer than 64 bytes in 
setCommonName adaptation algorithm. Also In the case the openssl fails to
read subject name from mimicking certificate does not set any subject to 
generated certification. (currently ssl_crtd crashes).