bug fix: assertion failed: client_side.cc:3720: "alg && param"
Someone can hit this assertion when have two "sslproxy_cert_adapt setCommonName"
configuration lines the first with parameter but the second without parameter:
  sslproxy_cert_adapt setCommonName{toThisOne} AN_IP
  sslproxy_cert_adapt setCommonName AN_IP

Inside ConnStateData::buildSslCertGenerationParams method inside the loop
  for (sslproxy_cert_adapt *ca = Config.ssl_client.cert_adapt; ....) {...}
the second time the loop entered the param is NULL and never set because 
certProperties.setCommonName is already set, so we hit the assertion.

The problem solved with removing the assertion.