Windows: portability layer updates for MinGW environment
* Move the util.h definitions into os/mswindows.h compat header.
* Add wrapper definitions for many MinGW specific POSIX symbols
* UPdate POSIX wrapper types

A few of these changes are suspect as to whether they work for CygWin and
MSCV. Particularly the POSIX wrapper type changes. We still need a set of
comprehensive MinGW, Cygwin and MSCV build tests to be done. These changes
were enough to build on MinGW some time back, but additional changes are
now likely to be needed and no run-time testing has been done so far.