Windows: revert broken attempt at removing #define macro wrappers
"using Squid::foo" style of defining when namespace functions to use when
there is a clash appears not to work on MinGW. We are forced to retain
the old #define macros.
This resolves "error: call of overloaded ... is ambiguous"

Update a few functions parameters to POSIX compliant APIs using socklen_t.
This resolves "error: invalid conversion from 'socklen_t*'"