Bug 3698: Add missing include to xstrerror.h
later versions were fixed as poart of trunk rev.12290 which added a lot
of errno.h includes. However all that is needed is this one to pull it
in through libcompat headers which cover all .cc.

xstrerror.h is the right place to add the include because it is the place
presenting a macro inlining the dependency.