Bug 3111: Mid-term fix for the forward.cc "err" assertion.
The assert is triggered when a close handler for the server connection
destroys FwdState before we have received anything from the origin 
server.
Instead of asserting, we now respond with a 502 (Bad Gateway) 
ERR_READ_ERROR.

That error seems the most appropriate single choice among available 
ones, but
it may be misleading (in access.log) when the close handler was called 
due to
client problems. Hopefully, another error will be logged in most of 
those
cases.