Fix lots of SNMP alignment, sizeof(), and memory uninitialized erorrs
They all stem from a few uses of "sizeof(name)" where name is a pointer
to an OID object rather than the object itself. This code is present as
far back as squid-2.6 and probably a lot further.

It appears to not have been causing obvious problems due to the
sizeof(oid*) [4 or 8 octets] on a pointer being larger than the
sizeof(oid) [1 or 2 octets] on the object stored into the allocated
memory.

 Detected by Coverity Scan. Issues 740305, 740306, 740306, 740307,
    740308, 740459, 740460, 740461, 740462, 740465, 740466, 740467,
    740468, 740469, 740470, 740471, 740472.