Polish: Drop redundant HttpMsgPointerT template
Despite the object and API changes being rather extensive
there are no logic changes in this patch.

The HttpMsgPointerT template was used to ref-count HttpMsg
objects and its children. It provided an API identical to
the RefCount API plus a few extensions for handling
polymorphism cast operations in the background.

This patch:

* update HttpMsg class hierarchy to use Lock API directly
 removing the _lock() and _unlock() virtual functions.

* update HttpControlMsg to use a HttpReply::Pointer directly
 since this type is available to that class a re-definition
 using HttpMsg* casting was not needed.

* update HTTPMSGLOCK() macro not to return a HttpMsg*
 The API polymorphism extensions provided by
 HttpMsgPointerT<> were only necessary to handle the
 HttpMsg* object casting this macro made a requirement.
 Once that assignment casting is removed the entire
 API extensions are no longer used.